OmniParser/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
einops==0.8.0
filelock==3.16.1
flash_attn==2.6.3
Flask==3.0.3
fsspec==2024.9.0
huggingface-hub==0.25.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
psutil==6.0.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.1.0
sympy==1.13.3
timm==1.0.9
tokenizers==0.19.1
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
